Fix MQTTClient unaligned access on ARM when using -Os
QoS value was an enum and this value was passed to MQTTSerialize_subscribe as an (int*).
So when -Os optimization is used the QoS enum was set to byte and placed at odd address.
But when the code reached MQTTSerialize_subscribe and the address of pointer to byte was casted to pointer to integer the LDR instruction caused HardFaul because of unaligned int access.

Fixed by forcing to pass char so now it should not matter which compiler is used.