Added Vendor Class Identifier in DHCP offers and requests.
The Vendor Class Identifier is a constant string that identifies the
kind of hardware that asks the DHCP server for a IP address. The ID is
passed via the option 60 of the DHCP messages.

The ID is set to "WIZnetSE" constant.

From a DHCP server point of view, this ID is useful for giving specific
configuration for Wiznet modules : IP address from different address
pool, different gateway, longer lease, etc.

Reference documentation :
https://datatracker.ietf.org/doc/html/rfc2132#section-9.13